// TuTurno.Backend/src/models/dtos/LocationDto.ts
import { OfficeDto } from "./OfficeDto";

export interface LocationDto {
    id: string;
    name: string;
    address: string;
    city: string;
    state: string;
    country: string;
    phone?: string | null;
    active?: boolean;
    createdAt?: Date;
    updatedAt?: Date;

    offices?: OfficeDto[];
}